Moved DigitalObjectSet into its own file to prevent cyclical import hell
