+++ src/app/digital/models/index.ts
@@ -2,6 +2,7 @@
 // This order is apparently extremely fragile so do not mess with it
 export * from "./DigitalCircuitDesigner";
 export * from "./DigitalComponent";
+export * from "./DigitalObjectSet";
 export * from "./ports";
 export * from "./DigitalWire";
 export * from "./Propagation";
