+++ src/app/digital/utils/ExpressionParser/index.ts
@@ -1,8 +1,9 @@
 import {Formats} from "./Constants/Formats";
 import {OperatorFormat} from "./Constants/DataStructures";
 
-import {DigitalComponent} from "digital/models/index";
-import {DigitalObjectSet} from "digital/utils/ComponentUtils";
+import {DigitalComponent} from "digital/models/DigitalComponent";
+import {DigitalObjectSet} from "digital/models/DigitalObjectSet";
+
 import {ValidateInputOutputTypes} from "./Utils";
 import {GenerateInputTree} from "./GenerateInputTree";
 import {GenerateTokens} from "./GenerateTokens";
